Move the PackageBuilder to the Spree namespace
Delete the calculator::weight namespace and move the PackageBuilder
to the default Spree::
